sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shield-sdk"
version = "0.1.0"
description = "Python SDK for the Shield AI rights management gateway"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shield_sdk*"]
